chore: release couplr 1.5.3
feat(memory): add memory_mode = "lazy" to match_couples(), compute_distances(), and assignment()

Computes each pairwise distance from the underlying feature data as the
solver needs it (method = "jv"/"auction", built-in distance metrics),
instead of allocating an n_left x n_right matrix up front. memory_mode =
"auto" estimates the dense matrix's footprint against free system RAM and
switches to lazy before a huge allocation. replace = TRUE, ratio > 1,
method = "greedy", custom distance functions, and full_match() are not
lazy-capable yet and error clearly.

fix(core): compute flat cost-matrix indices in 64-bit

lap::CostMatrix and several solvers' own raw index arithmetic (auction,
csa, cycle_cancel, munkres, ramshaw_tarjan, ssap_bucket, sap/ssp,
network_simplex) computed i * ncol + j in 32-bit int, silently wrapping
past ~46,341 square. All flat-index arithmetic is now 64-bit.

docs: correct memory-usage claims for greedy matching and calipers

build_cost_matrix() allocates the full matrix for every method including
"greedy"; couplr has no sparse matrix representation. Also documents
memory_mode = "lazy" on match_couples()/compute_distances(), which were
missing it despite already supporting it end to end.

fix(tests): update_constraints(max_distance = 0) now errors, matching
apply_max_distance()'s existing "must be positive" contract (already
enforced on the match_couples()/assignment() path) instead of the stale
permissive behavior update_constraints() had before this refactor.

+++ src/solvers/greedy_matching.cpp
@@ -32,8 +32,8 @@ struct CandidatePair {
 // Collect all valid pairs, sort by cost, greedily assign
 
 LapResult greedy_matching_sorted(const CostMatrix& cost, bool maximize) {
-    const int n = cost.nrow;
-    const int m = cost.ncol;
+    const int n = static_cast<int>(cost.nrow);
+    const int m = static_cast<int>(cost.ncol);
 
     // Handle empty case
     if (n == 0) {
@@ -115,8 +115,8 @@ LapResult greedy_matching_sorted(const CostMatrix& cost, bool maximize) {
 // For each unmatched row, find its best available column
 
 LapResult greedy_matching_row_best(const CostMatrix& cost, bool maximize) {
-    const int n = cost.nrow;
-    const int m = cost.ncol;
+    const int n = static_cast<int>(cost.nrow);
+    const int m = static_cast<int>(cost.ncol);
 
     // Handle empty case
     if (n == 0) {
@@ -180,8 +180,8 @@ LapResult greedy_matching_row_best(const CostMatrix& cost, bool maximize) {
 // Use priority queue for efficient selection (good for very large problems)
 
 LapResult greedy_matching_pq(const CostMatrix& cost, bool maximize) {
-    const int n = cost.nrow;
-    const int m = cost.ncol;
+    const int n = static_cast<int>(cost.nrow);
+    const int m = static_cast<int>(cost.ncol);
 
     // Handle empty case
     if (n == 0) {
