+++ src/solvers/solve_push_relabel.cpp
@@ -13,8 +13,8 @@
 namespace lap {
 
 LapResult solve_push_relabel(const CostMatrix& cost, bool maximize) {
-    const int n = cost.nrow;
-    const int m = cost.ncol;
+    const int n = static_cast<int>(cost.nrow);
+    const int m = static_cast<int>(cost.ncol);
 
     // Handle empty case
     if (n == 0) {
