+++ src/solvers/solve_sinkhorn.cpp
@@ -24,8 +24,8 @@ SinkhornResult solve_sinkhorn(
     const std::vector<double>& r_weights,
     const std::vector<double>& c_weights
 ) {
-    const int n = cost.nrow;
-    const int m = cost.ncol;
+    const int n = static_cast<int>(cost.nrow);
+    const int m = static_cast<int>(cost.ncol);
 
     // Handle empty case
     if (n == 0 || m == 0) {
